Abstracted styles
Even though it's less stress on the network to include the styles directly in the HTML file, there's no sense in the bad practice when such small amounts of data are being served anyways. I'd rather it be easier to modify styles on me as a developer than minutely faster for the small amount of end users.